Add the doxygen wrapper script.
The relative paths in the doxygen configuration file are relative to where
doxygen was run and not where the configuration file is located. This
causes a problem when trying to run doxygen using the meson, as we have no
guarantee of where it will run, thereby breaking the relative paths
in the configuration file in some situations. Example: Doxygen invocation
can work well with the vs2019 project, but it may be breaking in a ninja
project.

This script receives the path where doxygen command should be executed,
executes the command and returning to the original path.